Move body writers to separate file
